Refactor activity for GUI
The MainActivity now displays current settings for Approach and
Location values. Pressing the refresh button pulls approach
data from the arduino (not yet implemented) and calculates the
user's location. The send button sends the data to the server.

Note that we no longer send data on the start of the activity,
so that should be adjusted if some application relies in immediate
input.