Make approach and location strings fields
As fields (both strings), both are updated when the user presses
the refresh button. When posting data, both fields are referenced
when adding to the HTTP POST request.

Now, after the worker thread makes the POST request, it calls back
to the UI thread to post a toast message indicated the status of
the request.