Android now talks to arduino
This commit includes the usb-serial-driver, allowing the android
to talk to the arduino. The current issue is that the read() call
reads "as many bytes as possible" before a user-specified timeout,
rather than reading 4 bytes, the amount necessary for the int sent
by the arduino. The next step is to fix this issue.

Also, I have refactored the toast-printing logic into a function
which runs a toast on the UI thread. This is the defacto way of
debugging while the android cannot be hooked up to android studio
for logging.